Introduce mechanism to handle data updates on volumes.
When mounting container to volume, content is only copied initially when volume is empty. When updating datastories or geoscripting sandbox tutorial notebooks we want to be able to refresh the content on container startup without having the user manually remove all volumes (and thus loose ALL data on volumes). The new mechanism uses entrypoint scripts to copy the desired files from the image to the destination folder which in compose setup will be mounted to named volumes. Files which are not conflicting will not be removed, giving users the chance to persist changes easily by a copy and rename approach.