Persist jupyterlab settings & user python packages (#10)
Persist jupyterlab settings & user python packages.
This is achieved by using an override requirements.txt which is loaded upon container start in the entrypoint.sh. The packages are downloaded to a volume, which allows to have the available also in offline mode after an initial download. Care has been taken to clean the download folder when there is a new version of pip installed to avoid broken installation procedures.